LayoutTests/http/tests/media/video-buffered.html

<html>
<head>
</head>
<body onload="start()">
<p>Start playing a video with preloading enabled, do a seek near the
  end and check multiple buffered timeranges have been created.</p>
<video id="video" preload="auto" autobuffer></video>
<script src=video-test.js></script>
<script src=media-file.js></script>
<script>
    function delayedSeek() {
        testExpected("video.buffered.length", 1, "==");
        run("video.currentTime = video.duration - 0.5");
        waitForEvent("ended", ended);
    }

    function ended() {
        testExpected("video.buffered.length", 2, "==");
        endTest();
    }

    function start () {
        video = document.getElementById('video');
        video.src = "http://127.0.0.1:8000/media/video-throttled-load.cgi?nph=1&name=resources/counting-captioned.mov&throttle=20&type=video/quicktime";

        video.play();
        setTimeout(delayedSeek, 1000);
     }
</script>
</body>
</html>
